pyproject.toml
[tool.poetry]
name = "backendrepo"
version = "0.1.0"
description = ""
authors = ["sds7629 <sds7629@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
drf-spectacular = "^0.26.5"
django-debug-toolbar = "^4.2.0"
thefuzz = "^0.20.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
